requirements.txt
streamlit==1.29.0
psycopg2-binary
